+++ eruption-gui/src/ui/keyboard/hwdevices/null_keyboard.rs
@@ -31,6 +31,10 @@ impl NullKeyboard {
 }
 
 impl Keyboard for NullKeyboard {
+    fn get_device(&self) -> u64 {
+        0
+    }
+
     fn get_make_and_model(&self) -> (&'static str, &'static str) {
         ("Unknown", "Unknown")
     }