+++ eruption-gui/src/ui/mouse/hwdevices/roccat_kone_pure_ultra.rs
@@ -29,15 +29,21 @@ const BORDER: (f64, f64) = (32.0, 32.0);
 pub type Result<T> = std::result::Result<T, eyre::Error>;
 
 #[derive(Debug)]
-pub struct RoccatKonePureUltra {}
+pub struct RoccatKonePureUltra {
+    pub device: u64,
+}
 
 impl RoccatKonePureUltra {
-    pub fn new() -> Self {
-        RoccatKonePureUltra {}
+    pub fn new(device: u64) -> Self {
+        RoccatKonePureUltra { device }
     }
 }
 
 impl Mouse for RoccatKonePureUltra {
+    fn get_device(&self) -> u64 {
+        self.device
+    }
+
     fn get_make_and_model(&self) -> (&'static str, &'static str) {
         ("ROCCAT", "Kone Pure Ultra")
     }